fix issues with failing tests
- by avoid inline tests that mock a module
